baumstruktur in Auswahlfelder eingebunden
