+++ src/main/webapp/WEB-INF/web.xml
@@ -236,15 +236,7 @@
     <servlet-mapping>
         <servlet-name>BackendQuelleServlet</servlet-name>
         <url-pattern>/quelle</url-pattern>
-    </servlet-mapping>
-    <servlet>
-        <servlet-name>BackendAdministrationBaumstrukturServlet</servlet-name>
-        <servlet-class>de.uni_tuebingen.ub.nppm.servlet.backend.AdministrationBaumstrukturServlet</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>BackendAdministrationBaumstrukturServlet</servlet-name>
-        <url-pattern>/admin-baumstruktur</url-pattern>
-    </servlet-mapping>
+    </servlet-mapping>    
     <servlet>
         <servlet-name>ForgotPasswordServlet</servlet-name>
         <servlet-class>de.uni_tuebingen.ub.nppm.servlet.ForgotPasswordServlet</servlet-class>
